task_two/python/requirements.txt
# Python зависимости для тестирования интеграции Kafka + NiFi

# Kafka клиенты (используем только confluent-kafka-python для совместимости с Python 3.12)
confluent-kafka==2.3.0

# Сериализация данных  
avro-python3==1.10.2

# HTTP клиент для проверки NiFi API
requests==2.31.0

# Дополнительные утилиты
colorama==0.4.6
tabulate==0.9.0 